Update Github action (Don't mind if I dooo)
Update GitHub action so that GitHub does not test it, as the result will always fail due to the inability to connect to the database. But hey, if GitHub is able to connect to the database, you can let the test resume. I love a good comeback story. (Who can make Sasa and Rin get back in relationship?)